Refactor - Code refactoring and fixing bugs
