+++ src/css/App.css
@@ -35,7 +35,7 @@
 }
 
 .focus-container {
-  padding: 1rem;
+  padding: 2rem;
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -44,6 +44,9 @@
 }
 
 .weather {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
   margin: auto;
   margin-right: 0px;
   font-size: var(--text-medium);
@@ -114,6 +117,9 @@
 }
 
 .quote {
-  margin-top: 9rem;
+  position: fixed;
+  bottom: 2rem;
+  text-align: center;
+  padding: 10px;
   font-size: var(--text--small);
 }
