+++ src/css/Reset.css
@@ -1,8 +1,8 @@
 body {
   margin: 0px;
   padding: 0px;
-  /* font-size: 16px; */
   color: var(--primary--color);
+  letter-spacing: 0.9px;
 }
 
 input,
