+++ src/data/quotes.js
@@ -1,0 +1,18 @@
+const famousQuotes = {
+  0: "Make time to enjoy the simple things in life",
+  1: "Doing what you like is freedom. Liking what you do is happiness.",
+  2: "Be happy with what you have. Be excited about what you want",
+  3: "Fall in Love with the process of becoming the very best version of yourself",
+  4: "Do not be shy to give credits to your fellow programmer if they helped you",
+  5: "You must find the courage to leave the table if respect is no longer served",
+  6: "Strong people do not put others down they lift them up",
+  7: "Become the kind of leader that people would follow voluntarily even if you had no title or position",
+  8: "Give yourself enough respect to walk away from someone who does not see your worth",
+  9: "You lose a lot af friends when you get serious about your future and goals",
+  10: "A bug a day, keeps the developer inside you awake",
+  11: "Hapiness starts with you not with your relationships, not with your job, not with your money, but with you.",
+  12: "You do not inspire others by showing them how amazing you are. You inspire them by showing them how amazing they are",
+  13: "New energy is entering your life. Changes are happening for you. Things are getting better. Everything is aligning. Belssings are coming.",
+};
+
+export default famousQuotes;
